requirements.txt
pika
jsonschema
mysql-connector-python
